+++ src/types/soap.d.ts
@@ -1,5 +1,20 @@
+// src/types/soap.d.ts
 declare module "soap" {
-  const anySoap: any;
-  export default anySoap;
+  export interface IOptions {
+    [key: string]: any;
+  }
+
+  export interface ISecurity {
+    [key: string]: any;
+  }
+
+  export class Client {
+    [method: string]: any;
+  }
+
+  export function createClientAsync(
+    wsdlUrl: string,
+    options?: IOptions
+  ): Promise<Client>;
 }